Improve nearest neighbor and bin average code
