+++ tsdat/io/converters/_base_transform.py
@@ -1,4 +1,4 @@
-from typing import TYPE_CHECKING, Any, Dict, Hashable, List, Optional, Tuple
+from typing import List, Optional
 import xarray as xr
 
 from ...config.dataset import DatasetConfig  # pragma: no cover
